+++ src/mruby_integration/audio/music.cpp
@@ -29,15 +29,6 @@ mrb_set_music_pitch(mrb_state* mrb, mrb_value) -> mrb_value
   return mrb_nil_value();
 }
 
-auto
-mrb_get_music_time_length(mrb_state* mrb, mrb_value) -> mrb_value
-{
-  Music* music;
-  mrb_get_args(mrb, "d", &music, &Music_type);
-
-  return mrb_float_value(mrb, GetMusicTimeLength(*music));
-}
-
 void
 append_audio_music(mrb_state* mrb)
 {
@@ -51,9 +42,4 @@ append_audio_music(mrb_state* mrb)
                     "set_music_pitch",
                     mrb_set_music_pitch,
                     MRB_ARGS_REQ(2));
-  mrb_define_method(mrb,
-                    mrb->kernel_module,
-                    "get_music_time_length",
-                    mrb_get_music_time_length,
-                    MRB_ARGS_REQ(1));
 }