+++ CHANGELOG.md
@@ -97,6 +97,8 @@
 - Music#pause and Music#resume are now C++ backed
 - Removed pause_music_stream and resume_music_stream
 - Added cancel_main_loop
+- Music#length is now C++ backed
+- Removed get_music_time_length
 
 ## v0.3.14.1